==> multiple_promoter_transcript_level/01.summarize data at transcript level.ipynb <==
## 1. Data manipulate

The goal is to add a `transcript` column into overlap res. Each `transcript` has a unique `tss`. 

- use [foverlap](https://www.rdocumentation.org/packages/data.table/versions/1.11.8/topics/foverlaps) function to get `peak` overlapping with `TSS` regions.  

### 1.1 load all transcripts

require(data.table)
require(tidyverse)


#require(rtracklayer)

all.transcript <- fread('../dat/1910_v2/gencode.v19.1kb_all_possible_transcripts.bed',
                       col.names=c("seq",'start','end','gene'))
setkey(all.transcript,seq,start,end) # this step changed order 
all.transcript%>%head(1)
all.transcript%>%tail(1)
dim(all.transcript)

system('tail -n1 ../dat/1910_v2/gencode.v19.1kb_all_possible_transcripts.bed',intern = T)%>%cat

**fread changed order of rows (reordered based on seq) !!!**
- because set key 

saveRDS(all.transcript,'../dat/1910_v2/gencode.v19.1kb_all_possible_transcripts.Rds')

all.transcript<- readRDS('../dat/1910_v2/gencode.v19.1kb_all_possible_transcripts.Rds')

all.transcript%>%head(1)
all.transcript%>%tail(1)

all.transcript%>% rowid_to_column("tr.idx")%>%tail(1)

### 1.2 foverlap peaks/reads with TSS 

overlap.res <- fread('gzcat ../dat/1910_v2/islet.lf_mtx.gz')%>%
    separate(peak,into = c("seq","start","end"),sep = ":|-")%>%
    mutate(seq=paste0('chr',seq))
setDT(overlap.res)
overlap.res <-overlap.res[,c("start","end"):=lapply(.SD, as.integer), .SDcols = c("start","end")]
head(overlap.res,1)
dim(overlap.res)

####  foverlap only return first match tr and get dist vs prom peaks

foverlap.res <- data.table::foverlaps(x = overlap.res,y = all.transcript,
                 type="any", which=T,mult="first") #only choose 1st overlap
range(foverlap.res,na.rm = T)

sum(!is.na(foverlap.res))
sum(is.na(foverlap.res))
sum(!is.na(foverlap.res))+sum(is.na(foverlap.res)) 

overlap.res %>% mutate(tr.idx = foverlap.res) %>% left_join(all.transcript %>% 
    rowid_to_column("tr.idx"),by='tr.idx')%>%filter(!is.na(tr.idx))%>%head()

overlap.res <- overlap.res %>% mutate(tr.idx = foverlap.res) %>% left_join(all.transcript %>% 
    rowid_to_column("tr.idx") %>% select(tr.idx, gene), by = "tr.idx")
head(overlap.res, 1)
dim(overlap.res)

#### foverlap return all matched trs and get prom_peak-tr dic

foverlap.res <- data.table::foverlaps(x = overlap.res, y = all.transcript, type = "any", 
    which = T)  #only choose 1st overlap
foverlap.res%>%head(5)
range(foverlap.res$yid, na.rm = T)
sum(!is.na(foverlap.res$yid))
sum(is.na(foverlap.res$yid))
sum(!is.na(foverlap.res$yid)) + sum(is.na(foverlap.res$yid))

overlap.res%>%head(1)

peak_tr.dic <- overlap.res %>% rowid_to_column("xid") %>% left_join(foverlap.res) %>% 
    left_join(all.transcript %>% rowid_to_column("yid") %>% select(yid, gene))

peak_tr.dic%>%head(1)
peak_tr.dic%>%dim

peak_tr.dic <- peak_tr.dic %>% select(seq, start, end, yid, gene) %>% filter(!is.na(yid)) %>% 
    distinct %>% mutate(seq = sub("chr", "", seq)) %>% mutate(prom_peak = paste0(seq, 
    ":", start, "-", end)) %>% select(prom_peak, gene)
peak_tr.dic %>% head(1)
peak_tr.dic %>% dim
peak_tr.dic %>% pull(prom_peak) %>% unique %>% length

peak_tr.dic<- peak_tr.dic%>%distinct
peak_tr.dic%>%dim

#fwrite(peak_tr.dic,'../dat/1910_v2/peak_tr.dic.txt')
saveRDS(peak_tr.dic,'../dat/1910_v2/peak_tr.dic.rds')

## 5. filtering using umap's results

dat.all.cells <- fread('../dat/1910_v2/islet.cluster_labels.filt.txt')
head(dat.all.cells,1)
length(dat.all.cells$index)

overlap.res <- overlap.res%>% filter(cell %in% dat.all.cells$index)
dim(overlap.res)
length(unique(overlap.res$cell))

##  save

fwrite(overlap.res,'../dat/1910_v2/long_matrix_w_transcripts.txt')

### check idx

`(base) ➜  1901 git:(master) ✗ grep ABHD14A alpha_beta.promoter.long_matrix_w_transcripts_corrected.txt |awk -v FS=',' '{print $1,$2,$3,$4,$7}'| sort|uniq`

chr3 52007627 52009641 ABHD14A 92628

chr3 52007627 52009641 ABHD14A-ACY1 92628

`(base) ➜  1901 git:(master) ✗ grep ABHD14A alpha_beta.promoter.long_matrix_w_transcripts.txt |awk -v FS=',' '{print $1,$2,$3,$4,$7}'| sort|uniq`

chr3 52007627 52009641 ABHD14A 92628

chr3 52007627 52009641 ABHD14A-ACY1 92628

all.transcript[81744,]

system('sed -n 81744p ../dat/1910_v2/gencode.v19.1kb_all_possible_transcripts.bed',intern = T)%>%cat
