==> multiple_promoter_transcript_level/02.Binary_comp_prom.ipynb <==
require(data.table)
require(tidyverse)

## 1. Read data 

dat.mat <- fread("gzcat ../dat/1910_v2//long_matrix_w_transcripts.txt.gz") %>% filter(!is.na(tr.idx))
dim(dat.mat)
head(dat.mat)
# dat.mat$cell <- NULL #no need cell id

 Concepts: 
1. `promoter region`: -500bp + 500bp for all TSS in gencode 
2. `promoter Peaks`: peaks that overlap promoter region

dat.mat <- dat.mat %>% mutate(seq = sub("chr", "", seq)) %>% mutate(prom_peak = paste0(seq, 
    ":", start, "-", end)) %>% select(prom_peak,cell)%>%distinct()
dim(dat.mat)
dat.mat %>%head(1)

dat.mat <- dat.mat %>% left_join(fread("../dat/1910_v2/islet.cluster_labels.filt.txt") %>% 
    select(index, cluster_name), by = c(cell = "index"))
dat.mat %>% head(1)
dat.mat %>% dim

# get all cells 
dat.all.cells <- table((fread('../dat/1910_v2/islet.cluster_labels.filt.txt'))$cluster) # table for all cells
dat.all.cells
sum(dat.all.cells)
length(unique(dat.mat$cell))

## 2 Prepare data

# prepare data
setDT(dat.mat)
dat.mat <- dat.mat[, .N, by = .(prom_peak, cluster_name)] %>% arrange(prom_peak, 
    cluster_name)
head(dat.mat, 1)
dat.mat %>% nrow

dat.mat <- dat.mat%>%rename(cluster=cluster_name)

## 3. Fisher's exact test  between subtypes

As long as there is any  peaks in that cell, the peak is open in this cell. 

1. get total alpha 1 and alpha 2 cells 
2. test hit in alpha1 vs hit in alpha2  (create contentigen table)
3. perform [Fisher's exact test](https://en.wikipedia.org/wiki/Fisher%27s_exact_test) or [chi-squared test](https://en.wikipedia.org/wiki/Chi-squared_test)

Input: `dat.mat` and `dat.all.cells`
output: 

peak|n_celltype1|n_celltype2|total_cells|celltype | pval | frac_1 | frac_2| odds


### 3.1 perform fisher's exact test for all prom peaks

tr = "1:10002866-10003720"
celltypes <- c("alpha_1", "alpha_2")

test.dat <- dat.mat %>% filter(prom_peak == tr & cluster %in% celltypes)
test.dat
table.res<- test.dat$N; names(table.res) <- test.dat$cluster
# handle if 0 for one subtype
a = setdiff(celltypes, names(table.res))
table.res[a] <- 0
test.tab <- matrix(c(table.res[celltypes[1]], table.res[celltypes[2]], dat.all.cells[celltypes[1]] - 
    table.res[celltypes[1]], dat.all.cells[celltypes[2]] - table.res[celltypes[2]]), 
    byrow = T, nrow = 2, dimnames = list(expressed = c("Yes", "No"), subtype = celltypes))
test.tab
test.tab[1]
test.tab[2]


f1 <- test.tab[1]/(test.tab[1] + test.tab[2])
f2 <- test.tab[3]/(test.tab[3] + test.tab[4])
l <- ifelse(f1 > f2, "greater", "less")
test.res <- fisher.test(test.tab, alternative = l)
res <- list(pval = test.res$p.value, odds = test.res$estimate, type1_frac = f1, type2_frac = f2)
res

fisher.test(test.tab)


celltypes <- c('alpha_1','alpha_2')
dat.mat.sub <- dat.mat%>%filter(cluster %in% celltypes)
fun.ftestPerTr <- function(  tr = "1:10002866-10003720",
                             celltypes=c('alpha_1','alpha_2'),
                             dat=dat.mat.sub){
    

test.dat <- dat %>% filter(prom_peak == tr)
 table.res<- test.dat$N; names(table.res) <- test.dat$cluster

    # handle if 0 for one subtype 
    a=setdiff(celltypes,    names(table.res))
    table.res[a]<-0
    
    test.tab <- matrix(c(table.res[celltypes[1]], table.res[celltypes[2]], 
                         dat.all.cells[celltypes[1]]-table.res[celltypes[1]], 
                         dat.all.cells[celltypes[2]]-table.res[celltypes[2]]),
                       byrow =  T,       
                       nrow = 2,
                       dimnames = list(expressed = c("Yes", "No"),
                       subtype = celltypes))

    f1<-test.tab[1]/(test.tab[1]+test.tab[2])
    f2 <- test.tab[3]/(test.tab[3]+test.tab[4])
    l <-ifelse(f1>f2,'greater','less')
    test.res <- fisher.test(test.tab,alternative = l)
    res <- list(
                prom_peak =tr,
                pval=test.res$p.value,
                odds=test.res$estimate,
                type1_frac=f1,
                type2_frac=f2
               )
    res
}

##



system.time(fun.ftestPerTr(celltypes=celltypes))
t(fun.ftestPerTr(celltypes=celltypes))



# estimate time
celltypes <- list()
celltypes$alpha <- c("alpha_1", "alpha_2")
celltypes$beta <- c("beta_1", "beta_2")
celltypes$delta <- c("delta_1", "delta_2")
sum(sapply( c("alpha", "beta", "delta"),function(x) 
    length(unique((dat.mat%>% filter(cluster %in% celltypes[[x]]))$prom_peak))))*.029/3600

dat.mat%>%pull(prom_peak)%>%unique%>%length

celltypes <- list()
celltypes$alpha <- c("alpha_1", "alpha_2")
celltypes$beta <- c("beta_1", "beta_2")
celltypes$delta <- c("delta_1", "delta_2")

## test run

x<- "alpha"
dat.mat.sub <- dat.mat%>%filter(cluster %in% celltypes[[x]])
all.tr <- unique(dat.mat.sub$prom_peak)
length(all.tr)
do.call(rbind,lapply(all.tr[1:10],function(trr) t(fun.ftestPerTr(dat = dat.mat.sub, 
        tr = trr, celltypes = celltypes[[x]]))))

require(parallel)

res.transcript_level <- list()
# time consumming task
system.time(for (x in c("alpha", "beta", "delta")) {
    
    dat.mat.sub <- dat.mat %>% filter(cluster %in% celltypes[[x]])
    all.tr <- unique(dat.mat.sub$prom_peak)
    
    res.transcript_level[[x]] <- do.call(rbind, mclapply(all.tr, function(trr) t(fun.ftestPerTr(dat = dat.mat.sub, 
        tr = trr, celltypes = celltypes[[x]])), mc.cores = 10))
})

saveRDS(object = res.transcript_level, file = "../dat/1901/res.prom.peaks.fisher.Rds")

### 3.2 adding gene info

res.transcript_level <- rbind(data.frame(res.transcript_level$alpha,celltype="alpha"),
     data.frame(res.transcript_level$beta,celltype="beta"),
     data.frame(res.transcript_level$delta,celltype="delta"))
head(res.transcript_level)

peak_tr.dic <- readRDS('../dat/1910_v2/peak_tr.dic.rds')
peak_tr.dic%>%head(1)

head(res.transcript_level) %>% mutate(prom_peak = as.character(prom_peak)) %>% left_join(peak_tr.dic)

res.transcript_level <- res.transcript_level %>% mutate(prom_peak = as.character(prom_peak)) %>% 
    left_join(peak_tr.dic)


### 3.3 padj

# adjust for p value
m <- c("fdr", "bonferroni", "BY")
names(m) <- c("FDR", "padj.Bonferroni", "FDR.BY")
res.transcript_level.2 <- do.call(rbind, lapply(c("alpha", "beta", "delta"), function(ct) {
    res.genes_level.a.df <- res.transcript_level %>% filter(celltype == ct) %>% select(prom_peak, 
        pval, celltype) %>% distinct %>% as.data.frame
    res.genes_level.a.df %>% nrow %>% print
    
    for (i in 1:3) {
        res.genes_level.a.df[[names(m)[i]]] <- p.adjust(as.numeric(res.genes_level.a.df$pval), 
            method = m[i])
    }
    res.genes_level.a.df
}))
head(res.transcript_level.2, 1)
head(res.transcript_level.2 %>% filter(celltype == "beta"), 1)
head(res.transcript_level.2 %>% filter(celltype == "delta"), 1)

res.transcript_level %>% nrow
res.transcript_level.2 %>% nrow
res.transcript_level %>% left_join(res.transcript_level.2) %>% head(1)
res.transcript_level %>% left_join(res.transcript_level.2) %>% nrow
res.transcript_level <- res.transcript_level %>% left_join(res.transcript_level.2) %>% 
    select(prom_peak, gene, celltype, type1_frac, type2_frac, odds, pval, FDR, padj.Bonferroni, 
        FDR.BY)
fwrite(res.transcript_level, file = "../figures/Fig2/subfigs/fig_2.prom.sub_vs_sub.fisher.csv")

### 3.4 padj again after filtering peaks only present in peak calling for the tested cell types

res.transcript_level<-fread("../figures/Fig2/subfigs/fig_2.prom.sub_vs_sub.fisher.csv")
res.transcript_level%>%head(1)
res.transcript_level%>%nrow

res.transcript_level<-fread("../figures/Fig2/subfigs/fig_2.prom.sub_vs_sub.fisher.csv")
res.transcript_level%>%head(1)
res.transcript_level%>%nrow
res.transcript_level%>%filter(celltype=='delta')%>%pull(prom_peak)%>%unique%>%length

## double check if Only chose the peaks that found in cell types for delta
res.peak%>%head(1)
res.peak%>%filter(grepl('delta',V4))%>%head(1)
res.peak%>%filter(grepl('delta',V4))%>%tail(1)
res.peak%>%filter(grepl('delta',V4))%>%distinct%>%nrow
res.peak%>%filter(grepl('delta',V4))%>%pull(prom_peak)%>%unique%>%length

res.peak <- fread("../dat/1910_v2/islet.merged_peaks.anno.bed")
res.peak %>% head(1)
res.peak %>% nrow
res.peak  %>% mutate(V1 = sub("chr", "", V1)) %>% mutate(prom_peak = paste0(V1, 
    ":", V2, "-", V3))%>%filter(prom_peak %in% res.transcript_level$prom_peak)%>%dim
res.peak<- res.peak  %>% mutate(V1 = sub("chr", "", V1)) %>% mutate(prom_peak = paste0(V1, 
    ":", V2, "-", V3))%>%filter(prom_peak %in% res.transcript_level$prom_peak)%>%select(prom_peak,V4)
res.peak%>%head(1)


# adjust for p value
m <- c("fdr", "bonferroni", "BY")
names(m) <- c("FDR", "padj.Bonferroni", "FDR.BY")
res.transcript_level.2 <- do.call(rbind, lapply(c("alpha", "beta", "delta"), function(ct) {
    # ct <- 'alpha'
    res.genes_level.a.df <- res.transcript_level %>% filter(celltype == ct) %>% select(prom_peak, 
        pval, celltype) %>% distinct %>% as.data.frame
    res.genes_level.a.df %>% nrow %>% print
    res.genes_level.a.df <- res.genes_level.a.df %>% filter(prom_peak %in% (res.peak %>% 
        filter(grepl(ct, V4)) %>% pull(prom_peak)))
    res.genes_level.a.df %>% nrow %>% print
    
    for (i in 1:3) {
        res.genes_level.a.df[[names(m)[i]]] <- p.adjust(as.numeric(res.genes_level.a.df$pval), 
            method = m[i])
    }
    res.genes_level.a.df
}))
head(res.transcript_level.2, 1)
head(res.transcript_level.2 %>% filter(celltype == "beta"), 1)
head(res.transcript_level.2 %>% filter(celltype == "delta"), 1)

res.transcript_level %>% nrow
res.transcript_level.2 %>% nrow
res.transcript_level %>% head(1)
res.transcript_level.2 %>% left_join(res.transcript_level %>% select(prom_peak, celltype, 
    type1_frac, type2_frac, odds) %>% distinct) %>% nrow
res.transcript_level.2 %>% left_join(res.transcript_level %>% select(prom_peak, celltype, 
    type1_frac, type2_frac, odds) %>% distinct) %>% head(1)

res.transcript_level.2 %>% left_join(res.transcript_level %>% select(prom_peak, celltype, 
    type1_frac, type2_frac, gene, odds) %>% distinct) %>% nrow

res.transcript_level.2 %>% left_join(res.transcript_level %>% select(prom_peak, celltype, 
    type1_frac, type2_frac, gene, odds) %>% distinct) %>% head(1)


res.transcript_level.2 %>% left_join(res.transcript_level %>% select(prom_peak, celltype, 
    type1_frac, type2_frac, gene, odds) %>% distinct) %>% select(prom_peak, gene, 
    celltype, type1_frac, type2_frac, odds, pval, FDR, padj.Bonferroni, FDR.BY) %>% 
    head(1)

res.transcript_level <- res.transcript_level.2 %>% left_join(res.transcript_level %>% 
    select(prom_peak, celltype, type1_frac, type2_frac, gene, odds) %>% distinct) %>% 
    select(prom_peak, gene, celltype, type1_frac, type2_frac, odds, pval, FDR, padj.Bonferroni, 
        FDR.BY)

fwrite(res.transcript_level, file = "../figures/Fig2/subfigs/fig_2.prom.sub_vs_sub.fisher.csv")
