==> multiple_promoter_transcript_level/05a.Dynamical_data_prepare.ipynb <==
## link the promoter opening matrix with cell's pesudostates

require(data.table)
require(tidyverse)

### 1. load promoter's data 

from [summarize data at transcript level.ipynb](./summarize data at transcript level.ipynb)

dat.pro <- fread("gzcat ../dat/1910_v2//long_matrix_w_transcripts.txt.gz") %>% filter(!is.na(tr.idx))
dim(dat.pro)
head(dat.pro,1)
length(unique(dat.pro$cell))

dat.pro <- dat.pro %>% mutate(seq = sub("chr", "", seq)) %>% mutate(prom_peak = paste0(seq, 
    ":", start, "-", end)) %>% select(prom_peak, cell) %>% distinct()
dim(dat.pro)
length(unique(dat.pro$cell))
dat.pro %>% head(1)

### 2. load cell's pseduostates

dat.ps <- do.call(rbind, lapply(c("alpha", "beta", "delta"), function(x) fread(paste0("../dat/1910_v2/", 
    x, ".pseudotime.txt"), col.names = c("cell", "ps")) %>% mutate(celltype = x)))
head(dat.ps, 1)
tail(dat.ps, 1)
dim(dat.ps)
dat.ps %>% pull(celltype) %>% table
dat.ps %>% pull(celltype) %>% table %>% sum

dat.pro_ps <- inner_join(dat.pro,dat.ps)
dim(dat.pro_ps)
head(dat.pro_ps,1)
length(unique(dat.pro_ps$cell))

dat.pro_ps%>%select(cell,celltype)%>%distinct%>%pull(celltype)%>%table

fwrite(dat.pro_ps,"../dat/1910_v2/abd.promoter.long_matrix_w_transcripts_ps.txt")

## Determine bin number 

dat.pro_ps<- fread("../dat/1910_v2/abd.promoter.long_matrix_w_transcripts_ps.txt")
head(dat.pro_ps,1)
dim(dat.pro_ps)

ps.range <- lapply(c("alpha", "beta", "delta"), function(ct) dat.pro_ps %>% filter(celltype == 
    ct) %>% pull(ps) %>% range)
names(ps.range) <- c("alpha", "beta", "delta")
ps.range

examBinNumber <- function(N.bin = 51) {
    require('IRdisplay')
    ps.bks <- list()
    ps.bks$alpha <- seq(ps.range$alpha[1], ps.range$alpha[2] + 1e-05, length.out = N.bin)
    ps.bks$beta <- seq(ps.range$beta[1], ps.range$beta[2] + 1e-05, length.out = N.bin)
    ps.bks$delta <- seq(ps.range$delta[1], ps.range$delta[2] + 1e-05, length.out = N.bin)
    
    ps.bin.dat.sum <- rbind(dat.pro_ps %>% filter(celltype == "alpha") %>% mutate(ps_bin = cut(ps, 
        breaks = ps.bks$alpha, include.lowest = T, right = F)), dat.pro_ps %>% filter(celltype == 
        "beta") %>% mutate(ps_bin = cut(ps, breaks = ps.bks$beta, include.lowest = T, 
        right = F)), dat.pro_ps %>% filter(celltype == "delta") %>% mutate(ps_bin = cut(ps, 
        breaks = ps.bks$delta, include.lowest = T, right = F))) %>% select(-prom_peak, 
        -ps) %>% distinct()
    
    setDT(ps.bin.dat.sum)
    
    ps.bin.dat.sum <- ps.bin.dat.sum[, .(.N), by = .(ps_bin, celltype)]
    message(paste(N.bin, "# bins:\n"))
    message("quantiles of alpha cell number per bin:\n")
    t(quantile((ps.bin.dat.sum %>% filter(celltype == "alpha"))$N)) %>% display
    message("quantiles of beta cell number per bin:\n")
    t(quantile((ps.bin.dat.sum %>% filter(celltype == "beta"))$N)) %>% display
    message("quantiles of delta cell number per bin:\n")
    t(quantile((ps.bin.dat.sum %>% filter(celltype == "delta"))$N))
}

examBinNumber(101)

examBinNumber(51)

examBinNumber(21)
