==> multiple_promoter_transcript_level/05b.ps_100_bin_generate_dat.ipynb <==
require(data.table)
require(tidyverse)

## 1. Rationale

* due to the sparseness, the ps is binned to 100 (can change) 
* then chisq test wheter the ratio of openning is dependent on bins 

##  2. load data

dat.pro_ps<- fread("../dat/1910_v2/abd.promoter.long_matrix_w_transcripts_ps.txt")
head(dat.pro_ps,1)
tail(dat.pro_ps,1)
dim(dat.pro_ps)

## 3 bin ps to get fraction of state 1 cells

### 3.1 100 bins for a,b; 20 bins for d

ps.range <- lapply(c("alpha", "beta", "delta"), function(ct) dat.pro_ps %>% filter(celltype == 
    ct) %>% pull(ps) %>% range)
names(ps.range) <- c("alpha", "beta", "delta")
ps.range

N.bin <- 101
N.bin.d<- 21
ps.bks <- list()
ps.bks$alpha <- seq(ps.range$alpha[1], ps.range$alpha[2] + 1e-05, length.out = N.bin)
ps.bks$beta <- seq(ps.range$beta[1], ps.range$beta[2] + 1e-05, length.out = N.bin)
ps.bks$delta <- seq(ps.range$delta[1], ps.range$delta[2] + 1e-05, length.out = N.bin.d)

dat.pro_ps <- rbind(dat.pro_ps %>% filter(celltype == "alpha") %>% mutate(ps_bin = cut(ps, 
    breaks = ps.bks$alpha, include.lowest = T, right = F)) %>% mutate(ps_bin_idx = as.numeric(ps_bin)), 
    dat.pro_ps %>% filter(celltype == "beta") %>% mutate(ps_bin = cut(ps, breaks = ps.bks$beta, 
        include.lowest = T, right = F)) %>% mutate(ps_bin_idx = as.numeric(ps_bin)), 
    dat.pro_ps %>% filter(celltype == "delta") %>% mutate(ps_bin = cut(ps, breaks = ps.bks$delta, 
        include.lowest = T, right = F)) %>% mutate(ps_bin_idx = as.numeric(ps_bin))) %>% 
    arrange(celltype, ps_bin_idx)

dat.pro_ps %>% head(1)

fread('../dat/1910_v2/islet.cluster_labels.filt.txt')%>%head(1)
fread('../dat/1910_v2/islet.cluster_labels.filt.txt')%>%nrow

ps.bin.dat.sum <- dat.pro_ps %>% select(-prom_peak, -ps) %>% distinct() %>% select(-celltype) %>% 
    left_join(fread("../dat/1910_v2/islet.cluster_labels.filt.txt") %>% select(index, 
        cluster_name), by = c(cell = "index")) %>% separate(cluster_name, into = c("celltype", 
    "subtype"), sep = "_")
ps.bin.dat.sum %>% head(1)
ps.bin.dat.sum %>% nrow

setDT(ps.bin.dat.sum)
ps.bin.dat.sum <- ps.bin.dat.sum %>% unite("ps_bin", 3:2, sep = ":", )
ps.bin.dat.sum <- ps.bin.dat.sum[, .(.N), by = .(ps_bin, celltype, subtype)]
ps.bin.dat.sum <- (ps.bin.dat.sum %>% spread(key = "subtype", value = "N", fill = 0) %>% 
    mutate(N = (`1` + `2`), frac = `1`/(`1` + `2`)))
ps.bin.dat.sum %>% group_by(celltype) %>% summarise(tot = sum(N))
ps.bin.dat.sum %>% head(1)
ps.bin.dat.sum %>% nrow

ps.bin.dat.sum <- ps.bin.dat.sum %>% separate(ps_bin, into = c("ps_bin_idx", "ps_bin"), 
    sep = ":")
fwrite(ps.bin.dat.sum, "../dat/1910_v2/ps.bin.dat.sum.csv")

setDT(ps.bin.dat.sum)
ps.bin.dat.sum <- ps.bin.dat.sum[, .(.N), by = .(ps_bin, celltype)]
ps.bin.dat.sum%>%head(1)
ps.bin.dat.sum%>%nrow
cat(paste(N.bin, "# bins:\n"))
cat("quantiles of alpha cell number per bin:\n")
t(quantile((ps.bin.dat.sum %>% filter(celltype == "alpha"))$N))
cat("quantiles of beta cell number per bin:\n")
t(quantile((ps.bin.dat.sum %>% filter(celltype == "beta"))$N))
cat("quantiles of delta cell number per bin:\n")
t(quantile((ps.bin.dat.sum %>% filter(celltype == "delta"))$N))

### 3.2 Delta need bigger bin size
Or check `05a.Dynamical_data_prepare.ipynb`

N.bin <- 21
ps.bks <- list()
ps.bks$delta <- seq(ps.range$delta[1], ps.range$delta[2] + 1e-05, length.out = N.bin)

dat.pro_ps.d <- dat.pro_ps %>% filter(celltype == "delta") %>% mutate(ps_bin = cut(ps, 
    breaks = ps.bks$delta, include.lowest = T, right = F)) %>% mutate(ps_bin_idx = as.numeric(ps_bin)) %>% 
    arrange(celltype, ps_bin_idx)

dat.pro_ps.d %>% head(1)


ps.bin.dat.sum <- dat.pro_ps.d %>% select(-prom_peak, -ps) %>% distinct()
ps.bin.dat.sum %>% head(1)
ps.bin.dat.sum %>% nrow


setDT(ps.bin.dat.sum)
ps.bin.dat.sum <- ps.bin.dat.sum[, .(.N), by = .(ps_bin, celltype)]
ps.bin.dat.sum %>% head(1)
ps.bin.dat.sum %>% nrow
cat(paste(N.bin, "# bins:\n"))
cat("quantiles of delta cell number per bin:\n")
t(quantile((ps.bin.dat.sum %>% filter(celltype == "delta"))$N))
